+++ src/Driver/ReactDriver.php
@@ -57,27 +57,40 @@ class ReactDriver implements DriverInterface
         };
     }
 
+    public function defer(Closure $callback): mixed
+    {
+        return null;
+    }
+
     public function await(array &$stream): void
     {
-        $async = function ($ip, $fnFlows, $index, $then) {
-            $async = $this->async($fnFlows[$index]['job']);
+        $async = function (Closure $job) {
+            return function (mixed $data) use ($job) {
+                $async = $this->async($job);
 
-            if ($ip->data === null) {
-                $promise = $async();
-            } else {
-                $promise = $async($ip->data);
-            }
+                if ($data === null) {
+                    $promise = $async();
+                } else {
+                    $promise = $async($data);
+                }
 
-            $promise->then($then);
+                return static function ($then) use ($promise) {
+                    $promise->then($then);
+                };
+            };
         };
 
-        $loop = function () use (&$loop, &$stream, $async) {
+        $defer = static function (Closure $job) {};
+
+        $loop = function () use (&$loop, &$stream, $async, $defer) {
             $nextIp = null;
             do {
                 foreach ($stream['dispatchers'] as $index => $dispatcher) {
                     $nextIp = $dispatcher->dispatch(new PullEvent(), Event::PULL)->getIp();
                     if ($nextIp !== null) {
-                        $stream['dispatchers'][$index]->dispatch(new AsyncEvent($async, $nextIp, $stream['fnFlows'], $index, static function ($data) use (&$stream, $index, $nextIp) {
+                        $job = $stream['fnFlows'][$index]['job'];
+
+                        $stream['dispatchers'][$index]->dispatch(new AsyncEvent($async, $defer, $job, $nextIp, static function ($data) use (&$stream, $index, $nextIp) {
                             if ($data instanceof RuntimeException and array_key_exists($index, $stream['fnFlows']) && $stream['fnFlows'][$index]['errorJob'] !== null) {
                                 $stream['fnFlows'][$index]['errorJob']($data);
                             } elseif (array_key_exists($index + 1, $stream['fnFlows'])) {